PR feedback, fix tests
