+++ packages/template-retail-react-app/app/components/drawer-menu/index.test.js
@@ -7,10 +7,11 @@
 import React from 'react'
 import DrawerMenu from './index'
 import {renderWithProviders} from '../../utils/test-utils'
+import {mockCategories} from '../../mocks/mock-data'
 
 describe('DrawerMenu', () => {
     test('Renders DrawerMenu without errors', async () => {
-        renderWithProviders(<DrawerMenu isOpen={true} />)
+        renderWithProviders(<DrawerMenu isOpen={true} root={mockCategories.root} />)
 
         const drawer = document.querySelector('.chakra-portal')
         const accordion = document.querySelector('.chakra-accordion')
